refactor : types.js 파일 위치 수정
